pyproject.toml
[tool.pdm]
distribution = true

[tool.black]
line-length = 79
target-version = ["py310"]
skip-string-normalization = false
exclude = '''
/(
  \.git
  | \.hg
  | \.venv
  | venv
  | build
  | dist
  | src/modservice/grpc
)/
'''

[tool.isort]
profile = "black"
line_length = 79
known_first_party = ["app"]
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = [".venv", "src/modservice/grpc"]

[tool.flake8]
max-line-length = 120
extend-ignore = ["E203", "W503"]
per-file-ignores = [
    "__init__.py:F401",
    "tests/*:T20,ARG001"
]
exclude = [
  ".venv",
  "__pycache__",
  "build",
  "dist",
  "src/modservice/grpc",
]

[tool.ruff]
line-length = 120
target-version = "py310"
preview = true
exclude = [".venv", "build", "dist", "src/modservice/grpc"]

[tool.ruff.lint]
select = [
    "E",
    "F",
    "B",
    "UP",
    "SIM",
    "T20",
    "ARG",
    "C4",
    "RUF"
]
ignore = [
    "E203",
    "B008",
    "SIM105",
    "RUF001",
    "RUF100",
    "RUF003",
    "RUF002"
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T20", "ARG001"]

[tool.mypy]
python_version = "3.10"
strict = true
exclude = 'src(?:[/\\])modservice(?:[/\\])grpc'
warn_unused_configs = true
warn_return_any = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
explicit_package_bases = true
mypy_path = ["src"]
plugins = ["pydantic.mypy"]

[project]
name = "modservice"
version = "0.0.1"
description = "mod-service for esclient"
authors = [
    {name = "esclient"},
]
dependencies = [
    "grpcio-reflection<2.0.0,>=1.72.1",
    "pydantic<3.0.0,>=2.11.7",
    "pydantic-settings<3.0.0,>=2.10.1",
    "psycopg2-binary<3.0.0,>=2.9.10",
    "grpcio-tools<2.0.0,>=1.73.1",
    "watchfiles<2.0.0,>=1.1.0",
    "botocore>=1.34.0",
]
requires-python = "<4.0,>=3.10"
readme = "README.md"
license = {text = "MIT"}

[project.scripts]
run-server = "modservice.server:serve"

[dependency-groups]
dev = [
    "grpcio<2.0.0,>=1.72.1",
    "grpcio-tools<2.0.0,>=1.73.1",
    "protobuf<7.0.0,>=6.31.1",
    "pytest<9.0.0,>=8.4.1",
    "black>=23.1.0,<24.0.0",
    "isort>=5.12.0,<6.0.0",
    "flake8>=6.0.0,<7.0.0",
    "flake8-pyproject>=1.2.3",
    "ruff>=0.0.278,<0.1.0",
    "mypy>=1.5.1,<2.0.0",
    "types-protobuf>=6.30.2.20250703",
    "grpc-stubs>=1.53.0.6",
    "pytest-mock>=3.14.1",
    "types-psycopg2>=2.9.21.20250809",
]
